+++ packages/tasks-adapter-gcp/src/index.ts
@@ -1,3 +1,4 @@
+export type { CloudTaskConnectorConfig } from './CloudTaskConnector.js'
 export { CloudTaskConnector } from './CloudTaskConnector.js'
 export type { GCPServiceAccount } from './CloudTaskConnector.types.js'
 export { MockCloudTaskConnector } from './MockCloudTaskConnector.js'