+++ packages/tasks-adapter-hatchet/src/index.ts
@@ -1,1 +1,2 @@
+export type { HatchetConnectorConfig } from './HatchetConnector'
 export { HatchetConnector } from './HatchetConnector'